+++ src/components/Error.js
@@ -1,19 +1,11 @@
-import { useContext } from "react";
-import GameContext from "./GameContext";
+import Button from "./Button";
 
 const Error = () => {
-  const { game, setGame } = useContext(GameContext);
-
   return (
     <div className="container">
-      <h1 className="wrong">Error</h1>
-      <p>{game.error.message}</p>
-      <button
-        className="btn btn-primary"
-        onClick={() => setGame({ type: "RETURN_TO_MAIN_MENU" })}
-      >
-        Return To Main Menu
-      </button>
+      <h1 className="wrong">Oops</h1>
+      <p>The server might be down. Try again later.</p>
+      <Button content="Quit" />
     </div>
   );
 };
