+++ src/components/Result.js
@@ -1,47 +1,29 @@
 import { useContext } from "react";
-import GameContext from "./GameContext";
+import GameContext from "../contexts/game";
 import MusicPlayer from "./MusicPlayer";
+import Button from "./Button";
 
 const Result = () => {
-  const { game, setGame } = useContext(GameContext);
+  const { game } = useContext(GameContext);
 
-  if (game.guessWasCorrect) {
-    return (
-      <div className="container">
-        <h1>Score</h1>
-        <p>{game.score}</p>
+  return (
+    <div className="container">
+      <h1>Score</h1>
+      <p>{game.score}</p>
+      {game.guessWasCorrect ? (
         <h2 className="correct">Correct</h2>
-        <MusicPlayer />
-        <button
-          className="btn btn-primary"
-          onClick={() => setGame({ type: "START_NEXT_ROUND" })}
-        >
-          Start Next Round
-        </button>
-      </div>
-    );
-  } else {
-    return (
-      <div className="container">
-        <h1>Score</h1>
-        <p>{game.score}</p>
+      ) : (
         <h2 className="wrong">Game Over</h2>
-        <MusicPlayer />
-        <button
-          className="btn btn-primary"
-          onClick={() => setGame({ type: "REPLAY" })}
-        >
-          Replay
-        </button>
-        <button
-          className="btn btn-primary"
-          onClick={() => setGame({ type: "RETURN_TO_MAIN_MENU" })}
-        >
-          Return To Main Menu
-        </button>
-      </div>
-    );
-  }
+      )}
+      {game.guessWasCorrect ? (
+        <Button content="Continue" />
+      ) : (
+        <Button content="Replay" />
+      )}
+      <Button content="Quit" />
+      <MusicPlayer />
+    </div>
+  );
 };
 
 export default Result;
