+++ src/contexts/game.js
@@ -1,0 +1,5 @@
+import { createContext } from "react";
+
+const GameContext = createContext();
+
+export default GameContext;
